refactor: move the file move src/configtype.ts to src/@types/configtypes.d.ts
